// src/types.ts
export enum OrderSide {
  Buy = 0,
  Sell = 1,
}
// ...
export interface OpenSeaAccount {
  // Wallet address for this account
  address: string
  // Public configuration info, including "affiliate" for users who are in the OpenSea affiliate program
  config: string

  // This account's profile image - by default, randomly generated by the server
  profileImgUrl: string

  // More information explicitly set by this account's owner on OpenSea
  user: OpenSeaUser | null
}

export interface OpenSeaUser {
  // Username for this user
  username: string
}

/**
 * Simple, unannotated asset spec
 */
export interface Asset {
  // The asset's token ID, or null if ERC-20
  tokenId: string | null,
  // The asset's contract address
  tokenAddress: string,
  // The Wyvern schema name (e.g. "ERC721") for this asset
  schemaName?: WyvernSchemaName,
  // The token standard version of this asset
  version?: TokenStandardVersion,
  // Optional for ENS names
  name?: string,
  // Optional for fungible items
  decimals?: number
}

/**
 * Annotated asset contract with OpenSea metadata
 */
export interface OpenSeaAssetContract extends OpenSeaFees {
  // Name of the asset's contract
  name: string
  // Address of this contract
  address: string
  // Type of token (fungible/NFT)
  type: AssetContractType
  // Wyvern Schema Name for this contract
  schemaName: WyvernSchemaName

  // Total fee levied on sellers by this contract, in basis points
  sellerFeeBasisPoints: number
  // Total fee levied on buyers by this contract, in basis points
  buyerFeeBasisPoints: number

  // Description of the contract
  description: string
  // Contract's Etherscan / OpenSea symbol
  tokenSymbol: string
  // Image for the contract
  imageUrl: string
  // Object with stats about the contract
  stats?: object
  // Array of trait types for the contract
  traits?: object[]
  // Link to the contract's main website
  externalLink?: string
  // Link to the contract's wiki, if available
  wikiLink?: string
}

interface NumericalTraitStats {
  min: number,
  max: number
}

interface StringTraitStats {
  [key: string]: number,
}

/**
 * Annotated collection with OpenSea metadata
 */
export interface OpenSeaCollection extends OpenSeaFees {
  // Name of the collection
  name: string
  // Slug, used in URL
  slug: string
  // Accounts allowed to edit this collection
  editors: string[]
  // Whether this collection is hidden from the homepage
  hidden: boolean
  // Whether this collection is featured
  featured: boolean
  // Date collection was created
  createdDate: Date,

  // Description of the collection
  description: string
  // Image for the collection
  imageUrl: string
  // Image for the collection, large
  largeImageUrl: string
  // Image for the collection when featured
  featuredImageUrl: string
  // Object with stats about the collection
  stats: object
  // Data about displaying cards
  displayData: object,
  // Tokens allowed for this collection
  paymentTokens: OpenSeaFungibleToken[]
  // Address for dev fee payouts
  payoutAddress?: string,
  // Array of trait types for the collection
  traitStats: OpenSeaTraitStats,
  // Link to the collection's main website
  externalLink?: string
  // Link to the collection's wiki, if available
  wikiLink?: string
}

export interface OpenSeaTraitStats {
  [traitName: string]: NumericalTraitStats | StringTraitStats
}

/**
 * Annotated asset spec with OpenSea metadata
 */
export interface OpenSeaAsset extends Asset {
  assetContract: OpenSeaAssetContract
  collection: OpenSeaCollection
  // The asset's given name
  name: string
  // Description of the asset
  description: string
  // Owner of the asset
  owner: OpenSeaAccount
  // Orders on the asset. Null if asset was fetched in a list
  orders: Order[] | null
  // Buy orders (offers) on the asset. Null if asset in a list and didn't prefetch buy orders
  buyOrders: Order[] | null
  // Sell orders (auctions) on the asset. Null if asset in a list and didn't prefetch sell orders
  sellOrders: Order[] | null

  // Whether the asset is on a pre-sale (so token ids aren't real)
  isPresale: boolean
  // The cached and size-optimized image url for this token
  imageUrl: string
  // The image preview url for this token.
  // Note: Loses gif animation and may have issues with SVGs
  imagePreviewUrl: string
  // The original image url for this token
  imageUrlOriginal: string
  // Thumbnail url for this token
  imageUrlThumbnail: string
  // Link to token on OpenSea
  openseaLink: string
  // Link to token on dapp's site
  externalLink: string
  // Array of traits on this token
  traits: object[],
  // Number of times this token has been traded (sold)
  numSales: number
  // Data about the last time this token was sold
  lastSale: AssetEvent | null
  // The suggested background color for the image url
  backgroundColor: string | null,
  // The per-transfer fee, in base units, for this asset in its transfer method
  transferFee: BigNumber | string | null,
  // The transfer fee token for this asset in its transfer method
  transferFeePaymentToken: OpenSeaFungibleToken | null
}

/**
 * Defines a AssetEvent type which contains details about an event that occurred
 */
export interface AssetEvent {

  // The type of event
  eventType: AssetEventType

  // The timestamp of the transaction (if on-chain) or when the off-chain occurred
  eventTimestamp: Date

  // The auction type
  auctionType: AuctionType

  // The total price of the sale in the payment
  totalPrice: string

  // The transaction associated with the token sale
  transaction: Transaction | null

  // Details about the token used in the payment for this asset
  paymentToken: OpenSeaFungibleToken | null
}
// ...
export interface OpenSeaFungibleToken extends Token {
  imageUrl?: string
  ethPrice?: string
  usdPrice?: string
}

// Backwards compat
export type FungibleToken = OpenSeaFungibleToken

/**
 * Bundles of assets, grouped together into one OpenSea order
 * URLs for bundles are auto-generated from the name
 */
export interface OpenSeaAssetBundle {
  maker: OpenSeaAccount
  assets: OpenSeaAsset[]
  name: string
  slug: string
  permalink: string

  // Sell orders (auctions) on the bundle. Null if bundle in a list and didn't prefetch sell orders
  sellOrders: Order[] | null

  assetContract?: OpenSeaAssetContract
  description?: string
  externalLink?: string
}
// ...
export interface UnhashedOrder extends WyvernOrder {
  feeMethod: FeeMethod
  side: OrderSide
  saleKind: SaleKind
  howToCall: HowToCall
  quantity: BigNumber

  // OpenSea-specific
  makerReferrerFee: BigNumber
  waitingForBestCounterOrder: boolean
  englishAuctionReservePrice?: BigNumber

  metadata: ExchangeMetadata
}

export interface UnsignedOrder extends UnhashedOrder {
  hash: string
}

/**
 * Orders don't need to be signed if they're pre-approved
 * with a transaction on the contract to approveOrder_
 */
export interface Order extends UnsignedOrder, Partial<ECSignature> {
  // Read-only server-side appends
  createdTime?: BigNumber
  currentPrice?: BigNumber
  currentBounty?: BigNumber
  makerAccount?: OpenSeaAccount
  takerAccount?: OpenSeaAccount
  paymentTokenContract?: OpenSeaFungibleToken
  feeRecipientAccount?: OpenSeaAccount
  cancelledOrFinalized?: boolean
  markedInvalid?: boolean
  asset?: OpenSeaAsset
  assetBundle?: OpenSeaAssetBundle
}